* updates price type * adds reset dev db script * refactors product interface to its own file * adds products controller and show and create methods * adds tests for products get, create and create fail. * refactors products routes to use controller.

+++ migrations/sqls/20220309043741-products-up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE products (
     id SERIAL PRIMARY KEY,
     name varchar(100) NOT NULL,
-    price money DEFAULT 0.00,
+    price float(2) DEFAULT 0.00,
     category varchar(50)
 )
